requirements.txt
streamlit
pillow
streamlit-drawable-canvas
svgpathtools
transformers